requirements.txt
pytest
selenium
requests
allure-pytest
webdriver-manager